+++ tests/testthat/test-learn.R
@@ -42,6 +42,27 @@ test_that("Standard learning from MSSQL (schema 'dbo') or Postgres (schema 'publ
     dm_db_learned,
     dm_for_filter()[order_of_deletion]
   )
+
+  # learning without keys:
+  dm_db_learned_no_keys <- expect_silent(dm_from_src(src_db, learn_keys = FALSE))
+
+  # for learning from DB without learning the key relations
+  dm_for_filter_no_keys <- dm_for_filter()[order_of_deletion] %>%
+    dm_get_def() %>%
+    mutate(
+      pks = vctrs::list_of(new_pk()),
+      fks = vctrs::list_of(new_fk())) %>%
+    new_dm3()
+
+  # in case there happen to be other tables in schema "dbo" or "public"
+  dm_db_learned_no_keys <-
+    dm_db_learned_no_keys %>%
+    dm_select_tbl(!!!remote_tbl_names)
+
+  expect_equivalent_dm(
+    dm_db_learned_no_keys,
+    dm_for_filter_no_keys
+  )
 })
 
 
@@ -81,6 +102,22 @@ test_that("Learning from specific schema on MSSQL or Postgres works?", {
   expect_equivalent_dm(
     dm_db_learned,
     dm_for_disambiguate()[order_of_deletion]
+  )
+
+  # learning without keys:
+  dm_db_learned_no_keys <- expect_silent(dm_from_src(src_db, schema = schema_name, learn_keys = FALSE)) %>%
+    dm_select_tbl(!!!order_of_deletion)
+
+  dm_for_disambiguate_no_keys <- dm_for_disambiguate()[order_of_deletion] %>%
+    dm_get_def() %>%
+    mutate(
+      pks = vctrs::list_of(new_pk()),
+      fks = vctrs::list_of(new_fk())) %>%
+    new_dm3()
+
+  expect_equivalent_dm(
+    dm_db_learned_no_keys,
+    dm_for_disambiguate_no_keys
   )
 })