+++ NEWS.md
@@ -1,3 +1,20 @@
+# dm 0.1.10.9000
+
+- Internal changes only.
+
+
+# dm 0.1.10
+
+- Same as previous version.
+
+
+# dm 0.1.9.9001
+
+- Fix `dm_from_src()` for MSSQL when `learn_keys = FALSE` (#427).
+- Tests use `expect_snapshot()` everywhere (#456).
+- Fix compatibility with testthat 3.0.0 (#457).
+
+
 # dm 0.1.9.9000
 
 Internal changes only.